Modify bin file to support pdsh env

+++ setup.py
@@ -1,6 +1,6 @@
 import shutil
 import subprocess
-from numpy import require
+import os
 import pkg_resources
 import setuptools
 from setuptools.command.install import install as _install
@@ -9,7 +9,10 @@ from setuptools.command.install import install as _install
 def _post_install():
     p = subprocess.Popen('which deepspeed', stdout=subprocess.PIPE, shell=True)
     dp_bin_path = p.communicate()[0].decode('utf-8').strip()
+    if not os.path.isfile(dp_bin_path):
+        raise RuntimeError('deepspeed executable file not found, installation will stop...')
     shutil.copyfile('bin/deepspeed', dp_bin_path)
+    shutil.copyfile('bin/ds', os.path.join(dp_bin_path.split('/')[:-1]))
 
 class PostInstall(_install):
     def run(self):
@@ -19,7 +22,6 @@ class PostInstall(_install):
 required_dp_ver = '0.6.0'
 if pkg_resources.get_distribution("deepspeed").version != required_dp_ver:
     raise RuntimeError('deepspeed version should be {}, installation will stop...'.format(required_dp_ver))
-
 
 setuptools.setup(
     name="deepspeed_npu",
